Prologue, completed
This commit also corrects a few minor items in the
license and the readme file.

This scene must be the most complex of the whole
darn thing; don't expect to compile it (separately)
in less than 14 minutes (with 4GB of memory, that is).

It has quite a lot of fancy tricks, mostly in the
drum and piano parts.  Typesetting it took me three
whole days (and yet, I can't even imagine what it will
take to enter the last scene in Act 2).