Better functions
this commit (hopefully) solves once and for all the
piano staves bug.
Whenever a << \\ >> construct was used, the accidentals
didn't appear because the Voices that were
internally created didn't have the
Accidental_engraver. So I had to keep the engraver at
a Staff level, and just kill the accidentals on the fly
in the PseudoVoice.