Revert "Slight modification, upon request from V. Cortez"
This reverts commit 2ed55ff386319ae8014959ece5a34a399c5ad009.

I can't say I'm happy that despite my being willing to accept
some compromises, the singer is still not happy and just
doesn't want to sing a couple of measures, no matter what
notes I put in these. Well, so be it, but You Shall Not Pass
in *my* source code! Aaarrrr!!!