Quite a bunch of improvements
This commits adds several improvements and bugfixes unveiled
during the rehearsals here at Montpellier. The first violist
wondered why the notes were so high, when I had simply forgotten
a comma :-)

The LICENSE document is also updated and bumped to version .2,
which is the official version that covers this opera.