New piece: full text and skeleton
This is "étoile sans couleur", a rather lenghty piece
I'm writing for a stupid composition contest.
The rules are that you have to use a (long and somewhat
poorly written) poem, that must be spoken and not sung,
accompanied with at most four instruments -- one of which
must be a piano.
I've chosen a text by French poet Jacques Roubaud, who
used to be quite an important author here. As for the
instruments, I'm going all-acoustic with a marimba,
a classical guitar, and a Chinese hulusi flute (I'm
expecting the jury to be quite a bit snobish, so a
little exotism can't hurt here). The rest will be
your average "contemporary" piece, with lots of stuff
I don't normally use, such as microtones and
contemporary-looking notations. That's just for apparences,
otherwise there's isn't any actual artistic basis
-- long story short: contests are BS. But we all
knew that already, didn't we?