New piece: basic structure and overall movement types
This is (will be) a micro-suite in five very short movements
for harpsichord, written for Iowa-native soloist Richard Siegel.
Not sure about the title, I've been looking in vain for something
original and funny but have yet to find what I'm aiming for. (Oh,
and if the score could write itself too, that'd be swell.)