A few more notes (and date correction)
"Colchiques" is written in a slow 3/4 meter,
to convey a deliberate ancient (passacaglia)
feeling.  The melody is mostly written on
twelve-notes series comprised of classical
triads, thereby looking for a compromise
between total-chromatic language and familiar
musical elements.  (Which is also the point
of other pieces in the collection, for
example "Clotilde".)