Canzona, complete
I'm not sure this is worth anything (in fact, I'm pretty
convinced of the opposite).  Oh well.

One (mildly) amusing tidbit here: the lower hand is written
only in minor and major thirds, corresponding to the binary
encoding of Richard Siegel's name. (In 7-bit ASCII.)

(And, no, I'm not suggesting that it makes the music more
interesting or enjoyable.)