A few minor corrections.
Coming back from our first reading with
Thierry Barbé, I’ve noticed that a few
improvements were needed.  Interestingly,
they are mainly to be found in the very
first of these pieces, not in the (far
more complicated) pieces I’ve written
more recently and that required a quite
larger amount of time and work.